+++ __init__.py
@@ -22,7 +22,7 @@ bl_info = {
     "description": "Export objects in bundles",
     "author": "AquaticNightmare",
     "blender": (2, 82, 0),
-    "version": (2, 3, 3),
+    "version": (2, 3, 4),
     "category": "3D View",
     "location": "3D View > Tools Panel > Bundle Exporter",
     "warning": "",