+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.4](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_3_4)
+### Fixed
+- Export selected bundles was only exporting one bundle
+- Rare viewlayer error
+
+### Added
+- New Modifiers:
+    - Make normals consistent
+    - Apply Modifiers
+
 ## [2.3.3](https://gitlab.com/AquaticNightmare/bundle_exporter/-/releases/2_3_3)
 ### Fixed
 - Scene unit settings were being reset after each export